+++ src/sunflare/config.py
@@ -89,8 +89,65 @@ def _convert_widget_position_type(
     return x if isinstance(x, WidgetPositionTypes) else WidgetPositionTypes(x)
 
 
+@runtime_checkable
+class ModelInfoProtocol(AttrsInstance, Protocol):
+    """Protocol equivalent to :class:`~sunflare.config.ModelInfo`.
+
+    This protocol allows to implement the ``ModelInfo`` class
+    in packages that do not depend on ``sunflare`` directly.
+
+    The only required dependency is `attrs`_.
+
+    _attrs: https://www.attrs.org/en/stable/
+    """
+
+    plugin_name: str
+    plugin_id: str
+
+    def read_configuration(self, timestamp: float) -> dict[str, Any]:
+        """See :meth:`sunflare.config.ModelInfo.read_configuration`."""
+        ...
+
+    def describe_configuration(self, source: str) -> dict[str, Any]:
+        """See :meth:`sunflare.config.ModelInfo.describe_configuration`."""
+        ...
+
+
+@runtime_checkable
+class ControllerInfoProtocol(Protocol):
+    """Protocol equivalent to :class:`~sunflare.config.ControllerInfo`.
+
+    .. note::
+
+        This protocol is currently used only for type checking purposes.
+        within the Redsun application. In the future
+        we might be able to expose this for usage in
+        external packages.
+    """
+
+    plugin_name: str
+    plugin_id: str
+
+
+@runtime_checkable
+class WidgetInfoProtocol(Protocol):
+    """Protocol equivalent to :class:`~sunflare.config.WidgetInfo`.
+
+    .. note::
+
+        This protocol is currently used only for type checking purposes.
+        within the Redsun application. In the future
+        we might be able to expose this for usage in
+        external packages.
+    """
+
+    plugin_name: str
+    plugin_id: str
+    position: WidgetPositionTypes
+
+
 @define(kw_only=True)
-class WidgetInfo:
+class WidgetInfo(WidgetInfoProtocol):
     """Widget information model.
 
     All widget information models inherit from this class.
@@ -100,7 +157,10 @@ class WidgetInfo:
     plugin_name : ``str``, optional
         Widget plugin name.
         Equivalent to the name of the PyPI/Conda package.
-        Defaults to ``N/A``.
+    plugin_id : ``str``, optional
+        Widget plugin ID.
+        Associated with the exposed entry
+        point in the plugin manifest.
     repository : ``str``, optional
         Widget repository URL. Defaults to ``N/A``.
     position : ``WidgetPositionTypes``
@@ -108,10 +168,10 @@ class WidgetInfo:
     """
 
     plugin_name: str = field(
-        default="N/A", validator=validators.instance_of(str), on_setattr=setters.frozen
+        validator=validators.instance_of(str), on_setattr=setters.frozen
     )
-    repository: str = field(
-        default="N/A", validator=validators.instance_of(str), on_setattr=setters.frozen
+    plugin_id: str = field(
+        validator=validators.instance_of(str), on_setattr=setters.frozen
     )
     position: WidgetPositionTypes = field(
         converter=_convert_widget_position_type,
@@ -121,7 +181,7 @@ class WidgetInfo:
 
 
 @define(kw_only=True)
-class ControllerInfo:
+class ControllerInfo(ControllerInfoProtocol):
     """Controller information model.
 
     All controller information models inherit from this class.
@@ -131,58 +191,57 @@ class ControllerInfo:
     plugin_name : ``str``, optional
         Controller plugin name.
         Equivalent to the name of the PyPI/Conda package.
-        Defaults to ``N/A``.
-    repository : ``str``, optional
-        Controller repository URL. Defaults to ``N/A``.
+    plugin_id : ``str``, optional
+        Controller plugin ID.
+        Associated with the exposed entry point
+        in the plugin manifest.
     """
 
     plugin_name: str = field(
-        default="N/A", validator=validators.instance_of(str), on_setattr=setters.frozen
+        validator=validators.instance_of(str), on_setattr=setters.frozen
     )
-    repository: str = field(
-        default="N/A", validator=validators.instance_of(str), on_setattr=setters.frozen
+    plugin_id: str = field(
+        validator=validators.instance_of(str), on_setattr=setters.frozen
     )
 
 
 @define(kw_only=True)
-class ModelInfo:
+class ModelInfo(ModelInfoProtocol):
     """Base model for device information.
 
     All device information models inherit from this class.
 
     Attributes
     ----------
-    model_name : ``str``
-        Device model name.
-    vendor : ``str``, optional
-        Detector vendor.
-    serial_number : ``str``, optional
-        Detector serial number.
     plugin_name : ``str``, optional
         Model plugin name.
         Equivalent to the name of the PyPI/Conda package.
+    plugin_id : ``str``, optional
+        Model plugin ID.
+        Associated with the exposed entry point
+        in the plugin manifest.
+    vendor : ``str``, optional
+        Detector vendor.
         Defaults to ``N/A``.
-    repository : ``str``, optional
-        Model repository URL. Defaults to ``N/A``.
+    family : ``str``, optional
+        Detector family.
+        Defaults to ``N/A``.
     """
 
-    model_name: str = field(
+    plugin_name: str = field(
         validator=validators.instance_of(str), on_setattr=setters.frozen
     )
-    vendor: str = field(
-        default="N/A", validator=validators.instance_of(str), on_setattr=setters.frozen
+    plugin_id: str = field(
+        validator=validators.instance_of(str), on_setattr=setters.frozen
+    )
+    vendor: str = field(default="N/A", validator=validators.instance_of(str))
+    family: str = field(
+        default="N/A",
+        validator=validators.instance_of(str),
     )
     serial_number: str = field(
         default="N/A",
-        converter=str,
-        on_setattr=setters.frozen,
-    )
-
-    plugin_name: str = field(
-        default="N/A", validator=validators.instance_of(str), on_setattr=setters.frozen
-    )
-    repository: str = field(
-        default="N/A", validator=validators.instance_of(str), on_setattr=setters.frozen
+        validator=validators.instance_of(str),
     )
 
     __type_map = {
@@ -234,7 +293,7 @@ class ModelInfo:
             **{
                 key: {"value": value, "timestamp": timestamp}
                 for key, value in asdict(self).items()
-                if key != "model_name"
+                if key not in ["plugin_name", "plugin_id"]
             }
         }
 
@@ -267,36 +326,9 @@ class ModelInfo:
                     "shape": self.__get_shape(value),
                 }
                 for key, value in asdict(self).items()
-                if key != "model_name"
+                if key not in ["plugin_name", "plugin_id"]
             }
         }
-
-
-@runtime_checkable
-class ModelInfoProtocol(AttrsInstance, Protocol):
-    """Protocol equivalent to :class:`~sunflare.config.ModelInfo`.
-
-    This protocol allows to implement the ``ModelInfo`` class
-    in packages that do not depend on ``sunflare`` directly.
-
-    The only required dependency is `attrs`_.
-
-    _attrs: https://www.attrs.org/en/stable/
-    """
-
-    model_name: str
-    vendor: str
-    serial_number: str
-    plugin_name: str
-    repository: str
-
-    def read_configuration(self, timestamp: float) -> dict[str, Any]:
-        """See :meth:`sunflare.config.ModelInfo.read_configuration`."""
-        ...
-
-    def describe_configuration(self, source: str) -> dict[str, Any]:
-        """See :meth:`sunflare.config.ModelInfo.describe_configuration`."""
-        ...
 
 
 # helper private functions for type conversion
@@ -359,9 +391,9 @@ class RedSunSessionInfo:
         validator=validators.in_(FrontendTypes),
         on_setattr=setters.frozen,
     )
-    models: dict[str, ModelInfo] = field(factory=dict)
-    controllers: dict[str, ControllerInfo] = field(factory=dict)
-    widgets: dict[str, WidgetInfo] = field(factory=dict)
+    models: dict[str, ModelInfoProtocol] = field(factory=dict)
+    controllers: dict[str, ControllerInfoProtocol] = field(factory=dict)
+    widgets: dict[str, WidgetInfoProtocol] = field(factory=dict)
 
     @staticmethod
     def load_yaml(path: str) -> dict[str, Any]: