+++ src/sunflare/model/__init__.py
@@ -1,8 +1,10 @@
 from bluesky.protocols import check_supports
 
+from ._base import Model
 from ._protocols import ModelProtocol
 
 __all__ = [
     "ModelProtocol",
+    "Model",
     "check_supports",
 ]