+++ src/sunflare/model/_protocols.py
@@ -8,16 +8,15 @@ from typing_extensions import Protocol
 if TYPE_CHECKING:
     from typing import Any, Optional
 
-    from bluesky.protocols import Reading, SyncOrAsync
-    from event_model.documents.event_descriptor import DataKey
+    from bluesky.protocols import Descriptor, Reading, SyncOrAsync
 
-    from sunflare.config import ModelInfo
+    from sunflare.config import ModelInfoProtocol
 
 __all__ = ["ModelProtocol"]
 
 
 @runtime_checkable
-class ModelProtocol(Protocol):
+class ModelProtocol(Protocol):  # pragma: no cover
     """Minimal required protocol for a recognizable device in Redsun.
 
     Exposes the following Bluesky protocols:
@@ -26,37 +25,16 @@ class ModelProtocol(Protocol):
     - :class:`~bluesky.protocols.HasParent`
     - :class:`~bluesky.protocols.Configurable`
 
-    Additionally provides a `shutdown` method
-    to (optionally) perform cleanup operations.
+    Parameters
+    ----------
+    name : ``str``
+        Name of the model. Serves as a unique identifier for the object created from it.
+    model_info : :class:`~sunflare.config.ModelInfoProtocol`
+        Object implementing :class:`~sunflare.config.ModelInfoProtocol`.
     """
 
     @abstractmethod
-    def __init__(self, name: str, model_info: ModelInfo) -> None: ...
-
-    @abstractmethod
-    def configure(
-        self, *args: Any, **kwargs: Any
-    ) -> SyncOrAsync[tuple[Reading[Any], Reading[Any]]]:
-        """Configure the model.
-
-        The protocol allows to set new values for slow-changing parameters.
-
-        The method can be normal or `async`.
-
-        Parameters
-        ----------
-        *args
-            Positional arguments.
-        **kwargs
-            Keyword arguments.
-
-        Returns
-        -------
-        tuple[:class:`~bluesky.protocols.Reading`, :class:`~bluesky.protocols.Reading`]
-            A two-element tuple: the first element is the old configuration value;
-            the second element is the new configuration value set by the model.
-        """
-        ...
+    def __init__(self, name: str, model_info: ModelInfoProtocol) -> None: ...
 
     @abstractmethod
     def read_configuration(self) -> SyncOrAsync[dict[str, Reading[Any]]]:
@@ -74,7 +52,7 @@ class ModelProtocol(Protocol):
         ...
 
     @abstractmethod
-    def describe_configuration(self) -> SyncOrAsync[dict[str, DataKey]]:
+    def describe_configuration(self) -> SyncOrAsync[dict[str, Descriptor]]:
         """Describe the model configuration.
 
         Provides a description of each field of the model configuration.
@@ -83,16 +61,8 @@ class ModelProtocol(Protocol):
 
         Returns
         -------
-        dict[``str``, :class:`~event_model.DataKey`]
+        dict[``str``, :class:`~bluesky.protocols.Descriptor`]
             A dictionary with the description of each field of the model configuration.
-        """
-        ...
-
-    @abstractmethod
-    def shutdown(self) -> None:
-        """Perform cleanup operation of the model.
-
-        If not necessary, it can be left as a no-op.
         """
         ...
 
@@ -101,7 +71,7 @@ class ModelProtocol(Protocol):
     def name(self) -> str:
         """Model identifier.
 
-        Used to populate object_keys in the Event :class:`~event_model.DataKey`.
+        Used to populate object_keys in :class:`~event_model.DataKey`.
 
         See the following link for more information:
         https://blueskyproject.io/event-model/main/explanations/event-descriptors.html
@@ -120,7 +90,7 @@ class ModelProtocol(Protocol):
 
     @property
     @abstractmethod
-    def model_info(self) -> ModelInfo:
+    def model_info(self) -> ModelInfoProtocol:
         """The associated model information.
 
         It can return a subclass of :class:`~sunflare.config.ModelInfo`.