+++ tests/test_model.py
@@ -14,11 +14,13 @@ def test_detector_model(config_path: str) -> None:
     """Test the detector model info."""
     truth_dict: dict[str, Any] = {
         "First mock detector": {
-            "model_name": "MockDetectorModel",
+            "plugin_name": "mocks",
+            "plugin_id": "mock_detector",
             "sensor_size": [10, 10],
         },
         "Second mock detector": {
-            "model_name": "MockDetectorModel",
+            "plugin_name": "mocks",
+            "plugin_id": "mock_detector",
             "sensor_size": [10, 10],
             "exposure_egu": "s",
         },
@@ -40,18 +42,30 @@ def test_detector_model(config_path: str) -> None:
     for (name, cfg_info), (truth_name, truth_cfg_info) in zip(
         session.models.items(), truth_configs.items()
     ):
-        cfg = cast(MockDetectorInfo, cfg_info)
-        detector = MockDetector(name=name, cfg_info=cfg)
+        detector = MockDetector(name=name, cfg_info=cfg_info)
         assert isinstance(detector, ModelProtocol)
         assert detector.name == truth_name
         assert detector.parent is None
         assert detector.model_info == truth_cfg_info
 
+        reading = detector.read_configuration()
+        truth = {"sensor_size": {"value": (10, 10), "timestamp": 0}}
+        assert reading["sensor_size"]["value"] == truth["sensor_size"]["value"]
+        assert reading["sensor_size"]["timestamp"] == truth["sensor_size"]["timestamp"]
+        descriptor = detector.describe_configuration()
+        truth = {
+            "sensor_size": {"dtype": "array", "shape": [2], "source": "model_info"}
+        }
+        assert descriptor["sensor_size"]["dtype"] == truth["sensor_size"]["dtype"]
+        assert descriptor["sensor_size"]["shape"] == truth["sensor_size"]["shape"]
+        assert descriptor["sensor_size"]["source"] == truth["sensor_size"]["source"]
+
 
 def test_broken_detector_model() -> None:
     """Test the detector model info."""
     test_config: dict[str, Any] = {
-        "model_name": "MockDetectorModel",
+        "plugin_name": "mocks",
+        "plugin_id": "mock_detector",
         "sensor_size": [10, 10, 100],
         "pixel_size": [1, 1, 1],
     }
@@ -60,7 +74,8 @@ def test_broken_detector_model() -> None:
         MockDetectorInfo(**test_config)
 
     test_config = {
-        "model_name": "MockDetectorModel",
+        "plugin_name": "mocks",
+        "plugin_id": "mock_detector",
         "sensor_size": [0, 10],
         "pixel_size": [1, 1, 1],
     }
@@ -69,7 +84,8 @@ def test_broken_detector_model() -> None:
         MockDetectorInfo(**test_config)
 
     test_config = {
-        "model_name": "MockDetectorModel",
+        "plugin_name": "mocks",
+        "plugin_id": "mock_detector",
         "sensor_size": [10, 10],
         "pixel_size": [1, 1],
     }
@@ -78,7 +94,8 @@ def test_broken_detector_model() -> None:
         MockDetectorInfo(**test_config)
 
     test_config = {
-        "model_name": "MockDetectorModel",
+        "plugin_name": "mocks",
+        "plugin_id": "mock_detector",
         "sensor_size": [10, 10],
         "pixel_size": [0, 1, 1],
     }
@@ -91,11 +108,13 @@ def test_motor_model(config_path: str) -> None:
     """Test the motor model info."""
     truth_dict: dict[str, Any] = {
         "Single axis motor": {
-            "model_name": "MockMotorModel",
+            "plugin_name": "mocks",
+            "plugin_id": "mock_motor",
             "axes": ["X"],
         },
         "Double axis motor": {
-            "model_name": "MockMotorModel",
+            "plugin_name": "mocks",
+            "plugin_id": "mock_motor",
             "axes": ["X", "Y"],
             "step_egu": "mm",
         },
@@ -124,17 +143,25 @@ def test_motor_model(config_path: str) -> None:
 
 
 def test_broken_motor_model() -> None:
-    test_config: dict[str, Any] = {"model_name": "MockMotor", "axes": [1, 2]}
+    test_config: dict[str, Any] = {
+        "plugin_name": "mocks",
+        "plugin_id": "mock_motor",
+        "axes": [1, 2],
+    }
 
     with pytest.raises(ValueError):
         MockMotorInfo(**test_config)
 
-    test_config = {"model_name": "MockMotor", "axes": ["x", "a", "b"]}
+    test_config = {
+        "plugin_name": "mocks",
+        "plugin_id": "mock_motor",
+        "axes": ["x", "a", "b"],
+    }
 
     with pytest.raises(ValueError):
         MockMotorInfo(**test_config)
 
-    test_config = {"model_name": "MockMotor", "axes": []}
+    test_config = {"plugin_name": "mocks", "plugin_id": "mock_motor", "axes": []}
 
     with pytest.raises(ValueError):
         MockMotorInfo(**test_config)
@@ -144,20 +171,24 @@ def test_multi_model(config_path: str) -> None:
     """Test the multi model info."""
     truth_dict: dict[str, Any] = {
         "First mock detector": {
-            "model_name": "MockDetector",
+            "plugin_name": "mocks",
+            "plugin_id": "mock_detector",
             "sensor_size": [10, 10],
         },
         "Second mock detector": {
-            "model_name": "MockDetector",
+            "plugin_name": "mocks",
+            "plugin_id": "mock_detector",
             "sensor_size": [10, 10],
             "exposure_egu": "s",
         },
         "Single axis motor": {
-            "model_name": "MockMotor",
+            "plugin_name": "mocks",
+            "plugin_id": "mock_motor",
             "axes": ["X"],
         },
         "Double axis motor": {
-            "model_name": "MockMotor",
+            "plugin_name": "mocks",
+            "plugin_id": "mock_motor",
             "axes": ["X", "Y"],
             "step_egu": "mm",
         },
@@ -166,19 +197,19 @@ def test_multi_model(config_path: str) -> None:
     truth_config_detectors: dict[str, MockDetectorInfo] = {
         name: MockDetectorInfo(**cfg_info)
         for name, cfg_info in truth_dict.items()
-        if "Detector" in cfg_info["model_name"]
+        if "mock_detector" in cfg_info["plugin_id"]
     }
 
     truth_configs_motors: dict[str, MockMotorInfo] = {
         name: MockMotorInfo(**cfg_info)
         for name, cfg_info in truth_dict.items()
-        if "Motor" in cfg_info["model_name"]
+        if "mock_motor" in cfg_info["plugin_id"]
     }
 
     config_file = os.path.join(config_path, "multi_model_instance.yaml")
     config_dict = RedSunSessionInfo.load_yaml(config_file)
     for name, cfg_info in config_dict["models"].items():
-        if "Detector" in cfg_info["model_name"]:
+        if "mock_detector" in cfg_info["plugin_id"]:
             config_dict["models"][name] = MockDetectorInfo(**cfg_info)
         else:
             config_dict["models"][name] = MockMotorInfo(**cfg_info)
@@ -198,7 +229,8 @@ def test_multi_model(config_path: str) -> None:
             doc = detector.model_info.read_configuration()
             bs_config = {key: doc[key]["value"] for key in doc}
             truth = asdict(det_cfg_info)
-            del truth["model_name"]
+            truth.pop("plugin_name")
+            truth.pop("plugin_id")
             assert bs_config == truth
 
     for (name, mot_info), (truth_name, mot_cfg_info) in zip(
@@ -212,5 +244,6 @@ def test_multi_model(config_path: str) -> None:
             doc = motor.model_info.read_configuration()
             bs_config = {key: doc[key]["value"] for key in doc}
             truth = asdict(det_cfg_info)
-            del truth["model_name"]
+            truth.pop("plugin_name")
+            truth.pop("plugin_id")
             assert bs_config == truth